socket server extends Server class
